Hydra.Chain.Direct.MockServer

Description

This modules provide a mock server for the ouroboros mini-protocols. The mocking behavior is rather simple and only good to cover a few uses-cases.

In particular, the mock never errors, requests will always succeed with some value. Block productions in the chain-sync protocol is driven by the the tx-submission protocol. New transactions immediately causes a new singleton block to be created. The server maintains an in-memory database so that, multiple clients may connect to it, submit transactions and get a consistent state.

Some queries of the state-query protocol are also partially stubbed out, but the logic there is even dumber. GetUTxOByAddress will always yield an empty UTXO set, and GetLedgerTip returns an arbitrary tip, but always the same.

Documentation

withMockServer :: (NetworkId -> IOManager -> FilePath -> (ValidatedTx Era -> IO ()) -> IO a) -> IO a

A bracket-style resource acquisition for a mock server which responds to all three Ouroboros client mini-protocols. The callback includes the server fake network magic, the path to the server's socket and, a way to enqueue new transactions (and thus, force production of new blocks).