Hydra.Chain.DirectSpec

Description

Integration tests for the Direct chain component which does interact with a single mocked "node". For tests spinning up real cardano-nodes see EndToEndSpec.

XXX(AB): Are those tests really useful? What are they really testing? It seems to me we need either to turn those into proper Unit tests, eg. not requiring a full-blown mock server to run, or drop them altogether in favor of DirectChainSpec tests which test the real thing.

Documentation

spec :: Spec

mkSeedPayment :: NetworkId -> VerificationKey PaymentKey -> (ValidatedTx Era -> IO ()) -> IO ()

alice :: Party

bob :: Party

carol :: Party

data Log

Constructors

FromAlice DirectChainLog

FromBob DirectChainLog